Separate style defaults from CSS
