+++ svg_timeline/timeline_elements.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import Optional, Self
 
 from svg_timeline.geometry import Vector
-from svg_timeline.css import ClassNames
+from svg_timeline.svg_style_defaults import ClassNames
 from svg_timeline.notation import dt
 from svg_timeline.svg_primitives import Line, Text, Circle, Image, Rectangle, SvgGroup
 from svg_timeline.time_spacing import TimeSpacing
